+++ src/model/album.rs
@@ -74,9 +74,8 @@ pub(in crate) struct FullAlbums {
 /// Simplified Albums wrapped by Page object
 ///
 /// [Reference](https://developer.spotify.com/web-api/get-list-new-releases/)
-// TODO: Reduce this wrapper object to `Page<SimplifiedAlbum>`
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub struct PageSimpliedAlbums {
+#[derive(Deserialize)]
+pub (in crate) struct PageSimpliedAlbums {
     pub albums: Page<SimplifiedAlbum>,
 }